chore: fix remaining clippy warnings across workspace
- tui/app.rs: use is_none_or instead of map_or(true, ...)
- tui/agent_detail.rs: remove unnecessary u16 cast
- tui/input.rs: use enumerate().take() instead of needless range loop
- xtask/platform/mod.rs: allow dead_code on port_env field

+++ crates/tui/src/control/views/agent_detail.rs
@@ -124,7 +124,7 @@ fn render_input(f: &mut Frame, app: &mut App, area: Rect) {
     if app.input_mode {
         let (cursor_row, _) =
             crate::input::cursor_visual_pos(&app.input_buffer, app.input_cursor, inner_width);
-        let visible = area.height.saturating_sub(2) as u16;
+        let visible = area.height.saturating_sub(2);
         if cursor_row < app.input_scroll {
             app.input_scroll = cursor_row;
         }
